Refresh collections individually if no CTag found
This patch is necessary to get CardDAV with Google to work.
For some reason they do not return a ctag for the home set
and we have to fetch the individual collections instead.
This patch solves this by simply adding a fallback for when
the ctag is empty.

Summary: Fixes T9490

Reviewers: cmollekopf

Reviewed By: cmollekopf

Subscribers: kde-pim

Tags: #sink, #kde_pim

Maniphest Tasks: T9490

Differential Revision: https://phabricator.kde.org/D15055