Removed the etag cache.
The only thing it did is compare etags. The protocol does not
support getting only modified or deleted items, so we have to
diff based on etags anyways, which is easy enough to do in the client,
and not necessarily easier with the cache.

I thus don't think it makes sense to have it as part of a protocol
implementation.