Don't second guess the data we're giving
A failing sub-job should fail the entire thing.
The previous codepath of ignoring the error and trying to interpret it
as collection url would return an empty set of collections with some
servers. That means we could mask a network error and end up with an
empty set as supposedly successful result (Possibly what happened in the
error mentioned in the previous commit).